feat: clean up researcher service
